+++ CHANGELOG.md
@@ -3,6 +3,7 @@
     - NOTE: this is a breaking change. Tokens imported with previous app versions will not work and will need to be reimported.
 - Added a security step with PIN / Password entering requirement to disable app protection
 - Enhanced app protection page styling
+- Enhanced app scrollbar look
 
 # v0.6
 - Added open on startup feature